+++ src/main/java/leaguehub/leaguehubbackend/domain/participant/service/ParticipantService.java
@@ -2,43 +2,29 @@ package leaguehub.leaguehubbackend.domain.participant.service;
 
 import leaguehub.leaguehubbackend.domain.channel.dto.ParticipantChannelDto;
 import leaguehub.leaguehubbackend.domain.channel.entity.Channel;
-import leaguehub.leaguehubbackend.domain.channel.entity.ChannelRule;
-import leaguehub.leaguehubbackend.domain.channel.repository.ChannelRuleRepository;
 import leaguehub.leaguehubbackend.domain.channel.service.ChannelService;
 import leaguehub.leaguehubbackend.domain.email.exception.exception.UnauthorizedEmailException;
-import leaguehub.leaguehubbackend.domain.match.entity.MatchPlayerResultStatus;
-import leaguehub.leaguehubbackend.domain.match.entity.PlayerStatus;
-import leaguehub.leaguehubbackend.domain.match.repository.MatchPlayerRepository;
 import leaguehub.leaguehubbackend.domain.member.entity.BaseRole;
 import leaguehub.leaguehubbackend.domain.member.entity.Member;
-import leaguehub.leaguehubbackend.domain.member.exception.auth.exception.AuthInvalidTokenException;
-import leaguehub.leaguehubbackend.domain.member.repository.MemberRepository;
-import leaguehub.leaguehubbackend.domain.member.service.JwtService;
 import leaguehub.leaguehubbackend.domain.member.service.MemberService;
-import leaguehub.leaguehubbackend.domain.participant.dto.*;
-import leaguehub.leaguehubbackend.domain.participant.entity.GameTier;
 import leaguehub.leaguehubbackend.domain.participant.entity.Participant;
 import leaguehub.leaguehubbackend.domain.participant.entity.Role;
-import leaguehub.leaguehubbackend.domain.participant.exception.exception.*;
+import leaguehub.leaguehubbackend.domain.participant.exception.exception.InvalidParticipantAuthException;
+import leaguehub.leaguehubbackend.domain.participant.exception.exception.ParticipantNotFoundException;
+import leaguehub.leaguehubbackend.domain.participant.exception.exception.ParticipantRealPlayerIsMaxException;
 import leaguehub.leaguehubbackend.domain.participant.repository.ParticipantRepository;
-import leaguehub.leaguehubbackend.global.util.SecurityUtils;
 import lombok.RequiredArgsConstructor;
-import org.json.simple.parser.JSONParser;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
-import static leaguehub.leaguehubbackend.domain.match.entity.PlayerStatus.DISQUALIFICATION;
 import static leaguehub.leaguehubbackend.domain.member.entity.BaseRole.GUEST;
 import static leaguehub.leaguehubbackend.domain.member.entity.BaseRole.USER;
-import static leaguehub.leaguehubbackend.domain.participant.entity.RequestStatus.*;
-import static leaguehub.leaguehubbackend.domain.participant.entity.Role.*;
+import static leaguehub.leaguehubbackend.domain.participant.entity.RequestStatus.DONE;
+import static leaguehub.leaguehubbackend.domain.participant.entity.Role.HOST;
+import static leaguehub.leaguehubbackend.domain.participant.entity.Role.PLAYER;
 
 
 @Service
@@ -48,97 +34,9 @@ public class ParticipantService {
 
     private final ParticipantRepository participantRepository;
     private final MemberService memberService;
-    private final WebClient webClient;
-    private final JSONParser jsonParser;
     private final ChannelService channelService;
     @Value("${riot-api-key-1}")
     private String riot_api_key;
-    private final ChannelRuleRepository channelRuleRepository;
-    private final MatchPlayerRepository matchPlayerRepository;
-    private final JwtService jwtService;
-    private final MemberRepository memberRepository;
-
-    private final ParticipantWebClientService participantWebClientService;
-
-
-    //참여자의 권한 확인하는 서비스 - > 쿼리(조회해서 그냥 넘겨주기 떄문에)
-    public int findParticipantPermission(String channelLink) {
-        UserDetails userDetails = SecurityUtils.getAuthenticatedUser();
-        if (userDetails == null) return OBSERVER.getNum();
-        Member member = memberService.validateMember(userDetails.getUsername());
-
-
-        Optional<Participant> findParticipant = participantRepository.findParticipantByMemberIdAndChannel_ChannelLink(member.getId(), channelLink);
-
-        return findParticipant.map(participant -> participant.getRole().getNum())
-                .orElse(OBSERVER.getNum());
-    }
-
-    //해당 채널의 첫 번째 관리자를 찾는 서비스 - 쿼리
-    public String findChannelHost(String channelLink) {
-        return participantRepository.findParticipantByRoleAndChannel_ChannelLinkOrderById(HOST, channelLink).get(0).getNickname();
-    }
-
-
-    //채널 유저 조회 서비스  -> 쿼리
-
-    /**
-     * 해당 채널의 관전자인 유저들을 조회
-     *
-     * @param channelLink
-     * @return
-     */
-    public List<ResponseStatusPlayerDto> loadObserverPlayerList(String channelLink) {
-        Participant findParticipant = getParticipant(channelLink);
-
-        checkRole(findParticipant.getRole(), HOST);
-
-        List<Participant> findParticipants = participantRepository.findAllByChannel_ChannelLinkAndRoleAndRequestStatusOrderByNicknameAsc(channelLink, OBSERVER, NO_REQUEST);
-
-        return findParticipants.stream()
-                .filter(participant -> participant.getMember().getBaseRole() == USER)
-                .map(participant -> mapToResponseStatusPlayerDto(participant))
-                .collect(Collectors.toList());
-    }
-
-    //채널 유저 조회 서비스 --> 쿼리
-
-    /**
-     * 요청을 보낸 사람들을 조회
-     *
-     * @param channelLink
-     * @return
-     */
-    public List<ResponseStatusPlayerDto> loadRequestStatusPlayerList(String channelLink) {
-        Participant findParticipant = getParticipant(channelLink);
-        checkRole(findParticipant.getRole(), HOST);
-
-        List<Participant> findParticipants =
-                participantRepository.findAllByChannel_ChannelLinkAndRoleAndRequestStatusOrderByNicknameAsc
-                        (channelLink, OBSERVER, REQUEST);
-
-        return findParticipants.stream()
-                .map(participant -> mapToResponseStatusPlayerDto(participant))
-                .collect(Collectors.toList());
-    }
-
-    //채널 유저 조회 서비스 -->쿼리
-
-    /**
-     * 해당 채널의 PLAYER 역할인 유저들을 반환
-     *
-     * @param channelLink
-     * @return RequestPlayerDtoList
-     */
-
-    public List<ResponseStatusPlayerDto> loadPlayers(String channelLink) {
-
-        return participantRepository.findAllByChannel_ChannelLinkAndRoleAndRequestStatusOrderByNicknameAsc
-                        (channelLink, PLAYER, DONE)
-                .stream()
-                .map(participant -> mapToResponseStatusPlayerDto(participant))
-                .collect(Collectors.toList());
-    }
 
 
     //관리자가 요청 참가자 승인 서비스 -> 커맨드
@@ -184,40 +82,6 @@ public class ParticipantService {
     }
 
 
-    //관리자가 대회 참가자 실격 서비스 -> 커맨드
-    private ParticipantIdResponseDto disqualifiedToHost(String channelLink, ParticipantIdDto message) {
-        Participant myParticipant = findParticipantAccessToken(channelLink, message.getAccessToken());
-        checkRole(myParticipant.getRole(), HOST);
-
-        Participant findParticipant = getFindParticipant(channelLink, message.getParticipantId());
-
-        disqualificationParticipant(findParticipant);
-
-        return new ParticipantIdResponseDto(message.getMatchPlayerId(), DISQUALIFICATION.getStatus());
-    }
-
-    //대회 참가자가 기권 서비스 -> 커맨드
-    private ParticipantIdResponseDto selfDisqualified(String channelLink, ParticipantIdDto message) {
-        Participant myParticipant = findParticipantAccessToken(channelLink, message.getAccessToken());
-        checkRole(myParticipant.getRole(), PLAYER);
-
-        disqualificationParticipant(myParticipant);
-
-        return new ParticipantIdResponseDto(message.getMatchPlayerId(), DISQUALIFICATION.getStatus());
-    }
-
-    //기권 시키는 서비스 -> 커맨드
-    private void disqualificationParticipant(Participant findParticipant) {
-        findParticipant.disqualificationParticipant();
-        matchPlayerRepository.findMatchPlayersByParticipantId(findParticipant.getId())
-                .forEach(matchPlayer -> {
-                            matchPlayer.updatePlayerCheckInStatus(PlayerStatus.DISQUALIFICATION);
-                            matchPlayer.updateMatchPlayerResultStatus(MatchPlayerResultStatus.DISQUALIFICATION);
-                            matchPlayer.updateMatchPlayerScoreDisqualified();
-                        }
-                );
-    }
-
     //관리자가 관전자를 관리자로 권한 변경 서비스 -> 커맨드
 
     /**
@@ -234,40 +98,11 @@ public class ParticipantService {
         findParticipant.updateHostRole();
     }
 
-    //게임 카테고리에 따른 전적 검색 서비스 -> 쿼리
-
-    /**
-     * 게임 카테고리에 따라 요청 분할
-     *
-     * @param gameId
-     * @param category
-     * @return
-     */
-    public ResponseUserGameInfoDto selectGameCategory(String gameId, Integer category) {
-        ResponseUserGameInfoDto userGameInfoDto = new ResponseUserGameInfoDto();
-
-        if (category.equals(0)) {
-            userGameInfoDto = participantWebClientService.getTierAndPlayCount(gameId);
-        }
-
-        return userGameInfoDto;
-    }
-
 
     //해당 사용자가 호스트인지 확인 -> 커맨드
     public void checkAdminHost(String channelLink) {
         Participant participant = getParticipant(channelLink);
         checkRole(participant.getRole(), HOST);
-    }
-
-
-    //해당 사용자의 AccessToken 확인 서비스 - 커맨드(내부 동작)
-    private Participant findParticipantAccessToken(String channelLink, String accessToken) {
-        String personalId = jwtService.extractPersonalId(accessToken)
-                .orElseThrow(() -> new AuthInvalidTokenException());
-        Member member = memberRepository.findMemberByPersonalId(personalId).get();
-        Participant myParticipant = getParticipant(channelLink, member);
-        return myParticipant;
     }
 
 
@@ -277,8 +112,6 @@ public class ParticipantService {
             throw new ParticipantRealPlayerIsMaxException();
     }
 
-
-    //사용자의 이메일 인증인지 확인 - 내부 동작 여기 남는다
 
     /**
      * 이메일이 인증되었는지 확인
@@ -305,7 +138,6 @@ public class ParticipantService {
         return channelService.findParticipantChannelList();
     }
 
-    //
 
     /**
      * channelLink와 member로 해당 채널에서의 참가자 찾기 -> 쿼리
@@ -320,7 +152,6 @@ public class ParticipantService {
         return participant;
     }
 
-    //제 3자가 채널에서의 참가자 찾기
 
     /**
      * 제 3자가 channelLink와 participantId로 해당 채널에서의 참가자 찾기 --> 쿼리
@@ -352,8 +183,6 @@ public class ParticipantService {
     }
 
 
-    //자기 자신의 id를 가져오는 서비스 -> 내부동작 여기 남는다
-
     /**
      * 자기 자신이 participant를 찾을 때
      *
@@ -371,16 +200,6 @@ public class ParticipantService {
         return participant;
     }
 
-    //쿼리 컨트롤러에 반환할 dto를 정제하는 서비스 -> 쿼리 내부 메소드로 들어간다
-    private ResponseStatusPlayerDto mapToResponseStatusPlayerDto(Participant participant) {
-        ResponseStatusPlayerDto responsePlayerDto = new ResponseStatusPlayerDto();
-        responsePlayerDto.setPk(participant.getId());
-        responsePlayerDto.setNickname(participant.getNickname());
-        responsePlayerDto.setImgSrc(participant.getProfileImageUrl());
-        responsePlayerDto.setGameId(participant.getGameId());
-        responsePlayerDto.setTier(participant.getGameTier());
-        return responsePlayerDto;
-    }
 
     //해당 채널의 Rule이 맞는지 확인 -> 여기에 남는다
     public void checkRole(Role myRole, Role checkRole) {
